converts a string to a pointer

+++ 0x04-pointers_arrays_strings/100-atoi.c
@@ -12,8 +12,8 @@ int _atoi(char *s)
 	int i = 0;
 	int a = 1;
 	int b = 0;
-	unsigned int c = 1;
-	unsigned int v = 0;
+	long long int c = 1;
+	long long int v = 0;
 
 	while (s[i] < '0' || s[i] > '9')
 	{
